Thredds s3 cache is cleared when simultaneous downloads occur
